Create package so we can run the scripts.
