+++ analysis_and_sanity_checks/dataset_analysis/plot_energy_consistency_analysis.py
@@ -12,10 +12,9 @@ import logging
 import matplotlib.pyplot as plt
 import numpy as np
 from tqdm import tqdm
+from utilities import RESULTS_DIR
+from utilities.data import get_data_module
 
-from analysis_and_sanity_checks.dataset_analysis import RESULTS_DIR
-from analysis_and_sanity_checks.dataset_analysis.compute_dataset_covariance import \
-    get_data_module
 from diffusion_for_multi_scale_molecular_dynamics.analysis import (
     PLEASANT_FIG_SIZE, PLOT_STYLE_PATH)
 from diffusion_for_multi_scale_molecular_dynamics.callbacks.sampling_visualization_callback import \
@@ -32,11 +31,8 @@ plt.style.use(PLOT_STYLE_PATH)
 logger = logging.getLogger(__name__)
 setup_analysis_logger()
 
-dataset_name_3x3x3 = "Si_diffusion_3x3x3"
-dataset_name_2x2x2 = "Si_diffusion_2x2x2"
-dataset_name_1x1x1 = "Si_diffusion_1x1x1"
 
-list_dataset_names = [dataset_name_1x1x1, dataset_name_2x2x2, dataset_name_3x3x3]
+list_dataset_names = ["Si_diffusion_1x1x1", "Si_diffusion_2x2x2", "Si_diffusion_3x3x3"]
 
 if __name__ == "__main__":
