+++ analysis_and_sanity_checks/dataset_analysis/utilities/__init__.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+RESULTS_DIR = Path(__file__).parent.parent / "generated_artifacts"
+RESULTS_DIR.mkdir(parents=True, exist_ok=True)
